Move project into new package
