+++ flag_slurper/autopwn.py
@@ -10,7 +10,7 @@ from flag_slurper.autolib.models import SUDO_FLAG
 from . import utils, autolib
 from .autolib import models
 from flag_slurper.conf.config import Config
-from .project import Project
+from flag_slurper.conf.project import Project
 
 pass_config = click.make_pass_decorator(Config)
