+++ tests/test_autopwn.py
@@ -2,7 +2,7 @@ import pytest
 from click.testing import CliRunner
 
 from flag_slurper.cli import cli
-from flag_slurper.project import Project
+from flag_slurper.conf.project import Project
 
 
 @pytest.fixture
