+++ tests/test_credentials.py
@@ -4,7 +4,7 @@ from click.testing import CliRunner
 
 from flag_slurper.cli import cli
 from flag_slurper.autolib.models import CredentialBag, Credential
-from flag_slurper.project import Project
+from flag_slurper.conf.project import Project
 
 
 def test_add_credentials(db):
